requirements.txt
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.5
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
contourpy==1.3.1
cycler==0.11.0
faiss-cpu==1.7.4
Flask==2.3.2
fonttools==4.41.0
ftfy==6.1.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.24.1
packaging==23.1
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
regex==2023.6.3
requests==2.28.1
six==1.16.0
supabase==2.15.0
torch>=2.0.1
torchvision>=0.14.1
tqdm==4.65.0
typing_extensions>=4.4.0
urllib3==1.26.13
wcwidth==0.2.6
Werkzeug==2.3.6
zipp==3.16.2
